fix type of foreign key in migration files

+++ database/migrations/2022_03_15_083903_create_friends_table.php
@@ -15,7 +15,7 @@ class CreateFriendsTable extends Migration
     {
         Schema::create('friends', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('friend_id');
+            $table->unsignedBigInteger('friend_id');
             $table->timestamps();
             $table->softDeletes();
         });
