+++ database/migrations/2022_03_15_084259_create_votes_table.php
@@ -14,8 +14,8 @@ class CreateVotesTable extends Migration
     public function up()
     {
         Schema::create('votes', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('bookmark_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('bookmark_id');
             $table->timestamps();
         });
     }
