+++ database/migrations/2022_03_15_084544_create_group_user_table.php
@@ -15,8 +15,8 @@ class CreateGroupUserTable extends Migration
     {
         // GROUP_USER (group_id, user_id, created_at, subscribed_at, deleted_at)
         Schema::create('group_user', function (Blueprint $table) {
-            $table->unsignedInteger('group_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('group_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamp('subscribed_at');
             $table->timestamps();
             $table->softDeletes();
